chore: cleanup version group, invoke only once
