+++ packages/info/src/options.ts
@@ -5,4 +5,10 @@ export default [
         type: String,
         description: 'To get the output in specified format (accept json or markdown)',
     },
+    {
+        name: 'version',
+        usage: '--version',
+        type: Boolean,
+        description: 'Display the version of the package',
+    },
 ];
