+++ packages/webpack-cli/lib/utils/arg-parser.js
@@ -1,6 +1,6 @@
 const commander = require('commander');
 const logger = require('./logger');
-
+const { commands } = require('./cli-flags');
 const { defaultCommands } = require('./commands');
 
 /**
@@ -12,38 +12,28 @@ const { defaultCommands } = require('./commands');
  * @param {boolean} argsOnly false if all of process.argv has been provided, true if
  * args is only a subset of process.argv that removes the first couple elements
  */
-function argParser(options, args, argsOnly = false, name = '', helpFunction = undefined, versionFunction = undefined, commands) {
+function argParser(options, args, argsOnly = false, name = '', helpFunction) {
     const parser = new commander.Command();
     // Set parser name
     parser.name(name);
     parser.storeOptionsAsProperties(false);
 
-    if (commands) {
-        commands.reduce((parserInstance, cmd) => {
-            parser
-                .command(cmd.name)
-                .alias(cmd.alias)
-                .description(cmd.description)
-                .usage(cmd.usage)
-                .allowUnknownOption(true)
-                .action(async () => {
-                    const cliArgs = args.slice(args.indexOf(cmd.name) + 1 || args.indexOf(cmd.alias) + 1);
-                    return await require('../commands/ExternalCommand').run(defaultCommands[cmd.name], ...cliArgs);
-                });
-            return parser;
-        }, parser);
+    commands.reduce((parserInstance, cmd) => {
+        parser
+            .command(cmd.name)
+            .alias(cmd.alias)
+            .description(cmd.description)
+            .usage(cmd.usage)
+            .allowUnknownOption(true)
+            .action(async () => {
+                const cliArgs = args.slice(args.indexOf(cmd.name) + 1 || args.indexOf(cmd.alias) + 1);
+                return await require('../commands/ExternalCommand').run(defaultCommands[cmd.name], ...cliArgs);
+            });
+        return parser;
+    }, parser);
 
-        // Prevent default behavior
-        parser.on('command:*', () => {});
-    }
-
-    // Use customized version output if available
-    if (versionFunction) {
-        parser.on('option:version', () => {
-            versionFunction();
-            process.exit(0);
-        });
-    }
+    // Prevent default behavior
+    parser.on('command:*', () => {});
 
     // Use customized help output if available
     if (helpFunction) {
