+++ packages/webpack-cli/lib/utils/arg-utils.js
@@ -1,3 +1,5 @@
+const { commands } = require('./cli-flags');
+
 function hyphenToUpperCase(name) {
     if (!name) {
         return name;
@@ -18,6 +20,12 @@ function arrayToObject(arr) {
     }, {});
 }
 
+const isCommandUsed = (args) =>
+    commands.find((cmd) => {
+        return args.includes(cmd.name) || args.includes(cmd.alias);
+    });
+
 module.exports = {
     arrayToObject,
+    isCommandUsed,
 };
